refactor: price의 타입을 long에서 int로 변경

+++ gc-coffee/src/main/java/com/example/gccoffee/dto/product/ProductResponseDto.java
@@ -4,6 +4,6 @@ import com.example.gccoffee.model.Category;
 
 import java.util.UUID;
 
-public record ProductResponseDto(UUID productId, String productName, Category category, long price,
+public record ProductResponseDto(UUID productId, String productName, Category category, int price,
                                  String description) {
 }
